% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_bank.R
\docType{data}
\name{latvian_folksong_item_bank}
\alias{latvian_folksong_item_bank}
\title{Latvian folk song item bank}
\format{
Dataframe.
}
\usage{
latvian_folksong_item_bank
}
\description{
A 20 melodic phrases from Latvian folk songs.
}
\keyword{datasets}
