% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAAlatvian.R
\name{SAAlatvian}
\alias{SAAlatvian}
\title{The Latvian SAA}
\usage{
SAAlatvian(
  app_name = "SAAlatvian",
  no_folk_songs = 8,
  languages = c("lv", "en")
)
}
\arguments{
\item{languages}{}
}
\description{
The Latvian SAA
}
