% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAAlatvian.R
\name{SAAlatvian}
\alias{SAAlatvian}
\title{The Latvian SAA}
\usage{
SAAlatvian(
  app_name = "SAAlatvian",
  no_folk_songs = 10L,
  num_items = list(long_tones = 6L, arrhythmic = 0L, rhythmic = 10L),
  languages = c("lv", "en"),
  musicassessr_aws = FALSE
)
}
\arguments{
\item{musicassessr_aws}{}
}
\description{
The Latvian SAA
}
